% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_backend.R
\name{enable_JAX_backend}
\alias{enable_JAX_backend}
\alias{disable_JAX_backend}
\alias{status_JAX_backend}
\title{JAX backend}
\usage{
enable_JAX_backend(verbose = FALSE)

disable_JAX_backend(hard_unload = FALSE)

status_JAX_backend()
}
\arguments{
\item{verbose}{Whether to display informative messages any time the JAX backend is used by a \code{marginaleffects} function (defaults to \code{FALSE}).}

\item{hard_unload}{Whether to unload all Python libraries/functions (defaults to \code{FALSE}). Note that even after running \code{disable_JAX_backend(hard_unload = TRUE)}, the \code{reticulate} Python bindings will remain---see \url{https://github.com/rstudio/reticulate/issues/580#issuecomment-521364482}.}
}
\description{
Manages the JAX backend for \code{marginaleffects}.
\itemize{
\item \code{enable_JAX_backend()}: enables JAX backend.
\item \code{disable_JAX_backend()}: disables JAX backend, optionally unloading all internal Python libraries/functions.
\item \code{status_JAX_backend()}: checks the status of the JAX backend.
}
}
\examples{
library(marginaleffects)
library(marginaleffectsJAX)
enable_JAX_backend()
mod <- lm(mpg ~ hp + am, data = mtcars)
head(predictions(mod)) # using JAX
disable_JAX_backend()
head(predictions(mod)) # not using JAX

}
