% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_backend.R
\name{initialize_JAX_backend}
\alias{initialize_JAX_backend}
\alias{teardown_JAX_backend}
\title{JAX backend}
\usage{
initialize_JAX_backend()

teardown_JAX_backend(hard_unload = FALSE)
}
\arguments{
\item{hard_unload}{Whether to unload all Python libraries/functions (defaults to FALSE). Note that after running \code{teardown_JAX_backend(hard_unload = TRUE)}, reactivating the JAX backend takes more time.}
}
\description{
Manages the JAX backend for \code{marginaleffects}.
\itemize{
\item \code{initialize_JAX_backend()}: initializes JAX backend.
\item \code{teardown_JAX_backend()}: deactivates JAX backend, optionally unloading all internal Python libraries/functions.
}
}
\examples{
library(marginaleffects)
library(marginaleffectsJAX)
initialize_JAX_backend()
mod <- lm(mpg ~ hp + am, data = mtcars)
head(predictions(mod)) # using JAX
teardown_JAX_backend()
head(predictions(mod)) # not using JAX
}
