% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jax_jacobian.R
\name{jax_jacobian}
\alias{jax_jacobian}
\title{Compute Jacobian using JAX}
\usage{
jax_jacobian(coefs, newdata, model, hypothesis, calling_function, ...)
}
\description{
Compute Jacobian using JAX
}
\examples{
library(marginaleffects)
library(marginaleffectsJAX)
mod <- lm(mpg ~ hp + am, data = mtcars)
predictions(mod) |> head()
}
