% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_backend.R
\name{enable_JAX_backend}
\alias{enable_JAX_backend}
\alias{disable_JAX_backend}
\title{JAX backend}
\usage{
enable_JAX_backend()

disable_JAX_backend(hard_unload = FALSE)
}
\arguments{
\item{hard_unload}{Whether to unload all Python libraries/functions (defaults to FALSE). Note that after running \code{disable_JAX_backend(hard_unload = TRUE)}, reactivating the JAX backend takes more time.}
}
\description{
Manages the JAX backend for \code{marginaleffects}.
\itemize{
\item \code{enable_JAX_backend()}: enables JAX backend.
\item \code{disable_JAX_backend()}: disables JAX backend, optionally unloading all internal Python libraries/functions.
}
}
\examples{
library(marginaleffects)
library(marginaleffectsJAX)
enable_JAX_backend()
mod <- lm(mpg ~ hp + am, data = mtcars)
head(predictions(mod)) # using JAX
disable_JAX_backend()
head(predictions(mod)) # not using JAX
}
