% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manage_backend.R
\name{initialize_JAX_backend}
\alias{initialize_JAX_backend}
\title{JAX backend}
\usage{
initialize_JAX_backend()
}
\description{
Sets up a JAX backend for `marginaleffects`, using reticulate
}
\examples{
library(marginaleffects)
library(marginaleffectsJAX)
initialize_JAX_backend()
mod <- lm(mpg ~ hp + am, data = mtcars)
head(predictions(mod))
}
